@@ -1,14 +1,14 @@
 package com.sornambigai.biz.billgeneration.Contoller;
 
+import java.net.URI;
 import java.util.List;
 import java.util.Map;
 
+import com.sornambigai.biz.billgeneration.service.usercheck.model.ItemModel;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
 import com.sornambigai.biz.billgeneration.service.items.ItemsService;
 import com.sornambigai.entity.ItemCategoryMstEntity;
@@ -20,24 +20,25 @@ public class ItemsContoller {
 	@Autowired
 	private ItemsService itemsService;
 
-	@RequestMapping(value = "/getItems", method = RequestMethod.POST)
-	@ResponseBody
-	public List<ItemsEntity> getItems(@RequestBody Map<String, Object> requestMap) {
-		return itemsService.getFilteredItems(requestMap);
+	@GetMapping("/getItems")
+	public List<ItemsEntity> getItems(@RequestParam String itemCategoryId ) {
+		return itemsService.getFilteredItems(itemCategoryId);
 	}
 
-	@RequestMapping(value = "getItemCategories", method = RequestMethod.POST)
-	public List<ItemCategoryMstEntity> getItemCategories(@RequestBody Map<String, Object> requestMap) {
+	@GetMapping("/getItemCategories")
+	public List<ItemCategoryMstEntity> getItemCategories() {
 		return itemsService.getItemCategories();
 	}
 
 	@RequestMapping(value = "/addItem", method = RequestMethod.POST)
-	public boolean addItem(@RequestBody Map<String, String> requestMap) {
-		return itemsService.addItem(requestMap);
+	public ResponseEntity<ItemModel> addItem(@RequestBody Map<String, String> requestMap) {
+		ItemsEntity item = itemsService.addItem(requestMap);
+
+		return ResponseEntity.created(URI.create("/getItem/" + item.getItemId())).body(new ItemModel(item.getItemCategoryId(), item.getItemName()));
 	}
 
-	@RequestMapping(value = "/deleteItem", method = RequestMethod.POST)
-	public boolean deleteItem(@RequestBody Map<String, String> requestMap) {
-		return itemsService.deleteItem(requestMap);
+	@DeleteMapping(value = "/deleteItem/{id}")
+	public boolean deleteItem(@PathVariable String id) {
+		return itemsService.deleteItem(id);
 	}
 }
