@@ -1,10 +1,9 @@
 package com.sornambigai.entity;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
 
 import com.sornambigai.dto.ItemCategoryMstDto;
 
+import jakarta.persistence.Entity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
