@@ -1,9 +1,9 @@
 package com.sornambigai.entity;
 
-import javax.persistence.Entity;
 
 import com.sornambigai.dto.ItemsDto;
 
+import jakarta.persistence.Entity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
